Refs #9958: remove old discovery server from compilation
