+++ src/cpp/rtps/builtin/discovery/participant/timedevent/DSClientEvent.cpp
@@ -30,7 +30,7 @@
 
 
 namespace eprosima {
-namespace fastrtps {
+namespace fastdds {
 namespace rtps {
